Migrate orchestrator to bubus event bus with idiomatic handler-per-hook pattern
Replace the old synchronous generator-based executor.py with an event-driven
orchestrator using bubus EventBus. Each plugin hook is registered as its own
handler on the bus (keyed by CrawlEvent or SnapshotEvent), and the bus's
default serial handler execution ensures hooks run in registration order.

Side-effect cascades (Binary→Machine→config) chain through await bus.emit()
queue-jumps: when a handler awaits an emitted event, bubus processes it and
all its children synchronously before returning.

Event flow:
  await bus.emit(CrawlEvent) → each crawl hook handler fires serially
    → await bus.emit(ProcessEvent) → parses JSONL stdout
      → await bus.emit(BinaryEvent) → resolves via provider on_Binary hooks
        → await bus.emit(MachineEvent) → updates shared_config
  await bus.emit(SnapshotEvent) → each snapshot hook handler fires serially

- Add events.py: CrawlEvent, SnapshotEvent, ProcessEvent, BinaryEvent, MachineEvent
- Rewrite orchestrator.py: bus setup, per-hook handler registration, download()
- Delete executor.py (fully replaced by orchestrator.py)
- Update cli.py: asyncio.run(download(...)) with on_result callback
- Update tests: asyncio.run() wrapper for async download()
- Add bubus dependency to pyproject.toml
