+++ pyproject.toml
@@ -45,6 +45,7 @@ dependencies = [
     "platformdirs>=4.0.0",
     "requests>=2.28.0",
     "psutil>=7.2.1",
+    "bubus",
 ]
 
 [project.optional-dependencies]
@@ -117,6 +118,9 @@ reportMissingImports = true
 reportMissingTypeStubs = false
 pythonVersion = "3.11"
 pythonPlatform = "Linux"
+
+[tool.uv.sources]
+bubus = { git = "https://github.com/pirate/bbus" }
 
 
 [project.scripts]
